content/base/test/test_bug869000.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=869000
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 869000</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=869000">Mozilla Bug 869006</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 869000 **/

var c = new Text();
ok(c, "Text has been created without content");
is(c.data, "", "Text.data is ok");

c = new Text('foo');
ok(c, "Text has been created");
is(c.data, "foo", "Text.data is ok");

document.getElementById('display').appendChild(c);
ok(true, "Text has been added to the document");

</script>
</pre>
</body>
</html>
